+++ NEWS.md
@@ -7,6 +7,7 @@
 ### Major
 * Added `read_structure()` to read STRUCTURE formatted files. Added auto-read of .str files to `import.snpR.data()`.
 * Added `calc_abba_baba()` to do ABBA/BABA tests, including block jackknifing for significance.
+* Added `calc_diagnostic()` to plot basic diagnostic plots for snpRdata objects.
 
 ### Minor
 * Now redistributing part of Roy Francis's `pophelper` package, since it's not on CRAN. Switched to GPL license to allow this. The package is still cited automatically when generating a citation during `plot_structure()` or `plot_structure_map`(). Added a dependency on `stats` to allow for the re-packaging.
@@ -17,6 +18,7 @@
 
 ## Documentation
 * Removed the `snpR_association` vignette, since it required the `GMMAT`, `BGLR`, and `ranger` R packages to be installed, but they are only suggested, not required. This could cause vignette building to fail. May re-tool later to just use the internally implemented association tests.
+* Updated documentation for sfs related functions to more explicitly talk about ref and anc columns for folding.
 
 ## Bug fixes
 * Fixed a bug where an incorrect armitage stat would be calculated with `calc_association` if the major allele differed between the case and control.